MB-62178: Change encryption file header format
Make the header 64bytes fixed size instead of fixed and flex
length. This set a max id size of 36 bytes.

The motivation for making the entire header fixed size is
to simplify the logic. One may read the 64 first bytes of
a file and determine if it is encrypted or not (instead of
having to read first X bytes, then another X bytes and deal with
partial keys etc)

Change-Id: Ib7ae0896cb88eef36e00d0b444718986eee0c1fa
Reviewed-on: https://review.couchbase.org/c/platform/+/211336
Tested-by: Trond Norbye <trond.norbye@couchbase.com>
Reviewed-by: Pavlos Georgiou <pavlos.georgiou@couchbase.com>