MB-56147: Add EncryptionKey interface
We need an EncryptionKeyIface because a key can be a master key or a
derived key. Both have an associated cipher, id and the actual key.

A derived key has additional fields like label, context.

Magma will define a DerivedEncryptionKey that implements this interface
and composes over a cb::crypto::SharedEncryptionKey.

Change-Id: Ide3be815d3a18afb93be1e0f1e9456945daea2b3
Reviewed-on: https://review.couchbase.org/c/platform/+/212103
Reviewed-by: Pavlos Georgiou <pavlos.georgiou@couchbase.com>
Tested-by: Rohan Suri <rohan.suri@couchbase.com>