MB-63599: Add cbcat - Program to dump files [1/n]
Program used to cat potentially encrypted files. When
used on a system where everything is installed in the
default locations the tool may simply be used as

Usage:

    cbcat  audit-logs.*

(optionally specifying --password to specify the password
to decrypt secrets. Using '-' reads the password from
stdin).

When using on non-standard installations one must
provide path to gosecrets.cfg and dump-keys:

    cbcat --with-dump-keys=/path/to/dump-keys \
          --with-gosecrets=/path/to/gosecrets.cfg \
          --password -

The format for the output for dump-keys is specified
in a google document, but possibly not all various
output(s) it may produce which means there might be
error scenarios not covered. There is a number of
bug reports filed agains dump-keys for usability,
but this patch include a framework to query the
system for a key to be used from couchstore which
means it would be nice to get the framework
merged and fix internal usability later on).

Change-Id: I233b5d3f3929982dcc471ce123b7b6d1e7949353
Reviewed-on: https://review.couchbase.org/c/platform/+/215915
Reviewed-by: Vesko Karaganev <vesko.karaganev@couchbase.com>
Tested-by: Trond Norbye <trond.norbye@couchbase.com>