MB-63604: Read chunk by chunk
Let cbcat read the file chunk by chunk instead of using the
single read call. The motivation is that iff there is an error
in the file an exceptiion will be thrown and cause no output
to be generated and we want to display all chunks unitil we
hit the chunk containing an error.

Change-Id: I45c16ad3d39fc8e2e7d783e681dda8f88568a3c1
Reviewed-on: https://review.couchbase.org/c/platform/+/217443
Tested-by: Trond Norbye <trond.norbye@couchbase.com>
Reviewed-by: Pavlos Georgiou <pavlos.georgiou@couchbase.com>