Prefer using to_json for nlohmann::json over format_as
The cb::logger::Json format has a custom serialiser which converts
values to the desired format for logging.

It has a specialisation which applies to enums which are formattable via
fmt::format to make them use that, so we don't have to write many
to_json methods which just call format_as.

Unfortunately, we also have enums with have both format_as (formattable
via fmt::format) and to_json (for nlohmann::json) and they do the
opposite - serialise to json then to string for fmt. This is the case
for some flags enums, which print as "[A, B, C]" (in KV that is
DcpAddStreamFlag).

To handle that case, we have to either change all to_json for those
enums to be templated over the basic_json type so they work for our
cb::logger::Json type, or we can add a check here to default to the
to_json implementation for nlohmann::json (then convert to the
cb::logger::Json implicitly).

I've done the latter to minimise the need to refactor elsewhere.

Change-Id: I761a7be4e2deb14a810ed315374f7bda06fe1adc
Reviewed-on: https://review.couchbase.org/c/platform/+/219509
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>