MB-64508: Use the NoArena guard for the thread local buffer
The buffer has an initial size and a max size. The initial size
allocates a buffer when the thread_local is initialised and gets freed
when the thread_local is destroyed. Also, when the buffer grows (up to
max size), it allocates memory. Those cases are now under a NoArena
guard.

I've moved the buffer implementation out of the header, as it did not
need to be defined there.

Change-Id: Ice735f9ed3780ddab8cdd2b56605baf8cc81db3b
Reviewed-on: https://review.couchbase.org/c/platform/+/220524
Reviewed-by: Pavlos Georgiou <pavlos.georgiou@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>