MB-60861: Instantiate fmt::has_formatter conditionally
The JSON serializer defaults to calling fmt::to_string for enum values
which are formattable via fmtlib.

To do so, it checks fmt::has_formatter and passes to std::enable_if_t.
Because has_formatter is not used conditionally based on is_enum, the
compiler instantiates has_formatter for every T to check.

This is problematic, in the case where fmt/chrono.h is used in the same
compilation unit. The reason is because fmt/chrono.h will itself
specialise the formatting template used by fmtlib. Since we've already
instantiated that here for this check, we end up with:

  error: partial specialization of ‘struct
  fmt::v10::formatter<std::chrono::duration<_Rep1, _Period1>, Char>’
  after instantiation of ‘struct fmt::v10::formatter<
    std::chrono::duration<long int, std::ratio<1, 1000000000> >, ...>

So this patch restructures the check using fmt::has_formatter to only
happen once we've already established that T is an enum.

Change-Id: I44c8fea1e506d96c3af4357298a0ef5ce4ba476b
Reviewed-on: https://review.couchbase.org/c/platform/+/221766
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jim Walker <jim@couchbase.com>