MB-65536: Throw exception if mktemp fails to create files
cb::io::mktemp tries to create a unique filename based upon the
input filename and then tries to create the file. If an error
occurred while trying to create the filename we would generate
a new filename and try that until it succeeded.

This would cause an endless loop if opening the file failed for
different reasons than the file exists (for instance if the
caller don't have write access to the directory).

This patch change the behavior to throw an exception in these
situations.

Change-Id: Ifb9c5c38735e47225ec9d9c46895a9eb43ee0b56
Reviewed-on: https://review.couchbase.org/c/platform/+/224108
Reviewed-by: Rohan Suri <rohan.suri@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>