MB-65107: Clean up the jemalloc TSD on thread exit
On Windows, the C++ thread_local destructors are sequenced after the
jemalloc tls_callback hook which destroys the TSD state. Allocations we
do in thread_locals then could see a corrupt state. As part of this
ticket, we've patched our jemalloc fork, allowing it to re-create the
TSD state when that happens (instead of accessing the dangling TSD
pointer).

To ensure that a re-created TSD is always cleaned up after it has been
re-created in this edge case introduced by the jemalloc patch, we have
also added this new API je_thread_cleanup, which we need to call after
all C++ thread_local destructors have run.

See also:
 - https://github.com/couchbasedeps/jemalloc/commit/fc0eda6
 - https://github.com/jemalloc/jemalloc/issues/2825

Change-Id: I176bc3eb724430b0816cb8d342b5f458806a04f0
Reviewed-on: https://review.couchbase.org/c/platform/+/226005
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>
Reviewed-by: Jim Walker <jim@couchbase.com>
Tested-by: Vesko Karaganev <vesko.karaganev@couchbase.com>