MB-60861: Simplify cb::logger::Json type
The custom destruction logic was in an effort to avoid allocations at
all cost. The benchmark results indicate there is no difference.

The hack around the initializer_list_t constructor was moved to
kv_engine.

Also, we lose the ability to forward declare class Json (since it's a
template alias), but that was only used in the active/passive stream
headers and we can live without this forward decl.

The improvement is in lowered complexity. We only need to think of
cb::logger::Json now, never needing a separate cb::logger::BasicJsonType
type.

This change must be linked to a kv_engine change to replace
BasicJsonType with Json where it was used.

Change-Id: I15bf52ff5fa333644ab5b1f31e799556e6a6238c
Reviewed-on: https://review.couchbase.org/c/platform/+/229421
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>