MB-67520: Fix failure in murmurhash after narrowing fix
The murmurhash code does some strange things... which the narrowing
fixes broke.

This commit restores the core of murmurhash to its orginal code and
just changes the signature to use size_t for length.

The problem discovered is due to the very odd way that
MurmurHash3_x86_32 iterates the input data.

If we look at my previous change, getblock32 changed from

getblock32(const uint32_t* p, int i)

to

getblock32(const uint32_t* p, size_t i)

However the for loop is odd in that it uses a int i which is set
to a negative number.

    const auto* blocks = (const uint32_t*)(data + nblocks * 4);

    for (int i = -nblocks; i; i++) {

The value of i is then passed to getblocks32 and the code relies
on that negative i along with the blocks pointer being at the end
of the input to achieve a reverse loop.

We can address such oddities separatley, but for now let's just
focus on the narrowing warning which came from the function
signature use of int for the length of the input.

Change-Id: I244932a4870610d6b3c58c47457438aed3c3cc2b
Reviewed-on: https://review.couchbase.org/c/platform/+/230961
Reviewed-by: Vesko Karaganev <vesko.karaganev@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>