MB-69813: Don't check C++ standard on WIN32
The current version we're using of CMake expands CMAKE_CXX_STANDARD 23
to /std:c++latest due to the lack of a /std:c++23 at the time.

Recent versions of MSVC defines __cplusplus to 202400L instead of the
expected value of 202302L causing us to enter a code block where
we use

    #warning

Which is a syntax not supported by MSVC (it would prefer #pragma
message).  Given that we don't really need to run this check on all
platforms (its intention is to remind developers that we need to
verify the code when we upgrade the c++ standard) we can run this
check on a target which works as expected instead).

Change-Id: I5a046d52b9dc294e89c159220136ee0536e4d58f
Reviewed-on: https://review.couchbase.org/c/platform/+/237615
Reviewed-by: Blair Watt <blair.watt@couchbase.com>
Well-Formed: Restriction Checker
Tested-by: Build Bot <build@couchbase.com>