MB-69825: Add RW support for gzip
Extend cb::crypto::FileReader and FileWriter to read/write pure .gz
files when encryption isn't used. For file reade we'll treat it
as gzip if the file extension is .gz. For the file writer one needs
to specify GZIP as the compression (and no key). If a key is provided
we'll use ZLIB compression.

Change-Id: I55365061459422acce9f16cca1f7fbe1c5e48474
Reviewed-on: https://review.couchbase.org/c/platform/+/237683
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Mohammad Zaeem <mohammad.zaeem@couchbase.com>