MB-69825: Fix problems with ZlibStream and flush
It is a problem when using the BufferWriter and ZLibStream
that we might end up with multiple calls to flush() in a situation
where ZLib don't really want a flush. Its a bad idea to explicitly
create flush points for us as it would reduce the compression
ratio. Just bypass the explicit flush in this case

Change-Id: Id26f0ad1ba3a05bf565b1c95269d2f36e375c345
Reviewed-on: https://review.couchbase.org/c/platform/+/237698
Reviewed-by: Mohammad Zaeem <mohammad.zaeem@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>